+++ meta_package_manager/__init__.py
@@ -32,7 +32,7 @@ import logging
 from .bitbar import fix_environment
 
 
-__version__ = '2.4.1'
+__version__ = '2.5.0'
 """ Examples of valid version strings according :pep:`440#version-scheme`:
 
 .. code-block:: python